+++ lib/vscode/src/vs/ipc.d.ts
@@ -1,0 +1,1 @@
+../../../../typings/ipc.d.ts
