+++ lib/vscode/src/vs/server/entry.ts
@@ -1,10 +1,10 @@
 import { field } from '@coder/logger';
 import { setUnexpectedErrorHandler } from 'vs/base/common/errors';
-import { CodeServerMessage, VscodeMessage } from 'vs/server/ipc';
+import * as proxyAgent from 'vs/base/node/proxy_agent';
+import { CodeServerMessage, VscodeMessage } from 'vs/ipc';
 import { logger } from 'vs/server/node/logger';
 import { enableCustomMarketplace } from 'vs/server/node/marketplace';
 import { Vscode } from 'vs/server/node/server';
-import * as proxyAgent from 'vs/base/node/proxy_agent';
 
 setUnexpectedErrorHandler((error) => logger.warn(error instanceof Error ? error.message : error));
 enableCustomMarketplace();
