+++ lib/vscode/src/vs/server/node/server.ts
@@ -1,7 +1,7 @@
 import { field } from '@coder/logger';
-import { release } from 'os';
 import * as fs from 'fs';
 import * as net from 'net';
+import { release } from 'os';
 import * as path from 'path';
 import { Emitter } from 'vs/base/common/event';
 import { Schemas } from 'vs/base/common/network';
@@ -10,6 +10,7 @@ import { getMachineId } from 'vs/base/node/id';
 import { ClientConnectionEvent, IPCServer, IServerChannel, ProxyChannel } from 'vs/base/parts/ipc/common/ipc';
 import { LogsDataCleaner } from 'vs/code/electron-browser/sharedProcess/contrib/logsDataCleaner';
 import { main } from 'vs/code/node/cliProcessMain';
+import { Query, VscodeOptions, WorkbenchOptions } from 'vs/ipc';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { ConfigurationService } from 'vs/platform/configuration/common/configurationService';
 import { ExtensionHostDebugBroadcastChannel } from 'vs/platform/debug/common/extensionHostDebugIpc';
@@ -40,14 +41,13 @@ import { IRequestService } from 'vs/platform/request/common/request';
 import { RequestChannel } from 'vs/platform/request/common/requestIpc';
 import { RequestService } from 'vs/platform/request/node/requestService';
 import ErrorTelemetry from 'vs/platform/telemetry/browser/errorTelemetry';
+import { resolveCommonProperties } from 'vs/platform/telemetry/common/commonProperties';
 import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import { TelemetryLogAppender } from 'vs/platform/telemetry/common/telemetryLogAppender';
 import { TelemetryService } from 'vs/platform/telemetry/common/telemetryService';
 import { combinedAppender, NullTelemetryService } from 'vs/platform/telemetry/common/telemetryUtils';
 import { AppInsightsAppender } from 'vs/platform/telemetry/node/appInsightsAppender';
-import { resolveCommonProperties } from 'vs/platform/telemetry/common/commonProperties';
 import { TelemetryChannel } from 'vs/server/common/telemetry';
-import { Query, VscodeOptions, WorkbenchOptions } from 'vs/server/ipc';
 import { ExtensionEnvironmentChannel, FileProviderChannel, TerminalProviderChannel } from 'vs/server/node/channel';
 import { Connection, ExtensionHostConnection, ManagementConnection } from 'vs/server/node/connection';
 import { TelemetryClient } from 'vs/server/node/insights';
