+++ src/node/vscode.ts
@@ -2,7 +2,7 @@ import { logger } from "@coder/logger"
 import * as cp from "child_process"
 import * as net from "net"
 import * as path from "path"
-import * as ipc from "../../lib/vscode/src/vs/server/ipc"
+import * as ipc from "../../typings/ipc"
 import { arrayify, generateUuid } from "../common/util"
 import { rootPath } from "./constants"
 import { settings } from "./settings"
